% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{families}
\alias{families}
\alias{tests}
\alias{samples}
\alias{statistics}
\alias{parameters}
\title{Create list of character strings}
\usage{
families(...)
}
\arguments{
\item{\dots}{defines character strings to be passed into the function.}
}
\description{
This function is used mostly for user's convenience. It simply creates a
list of character strings.
}
\references{
\url{http://gpaux.github.io/Mediana/}
}
