% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{SampleSize}
\alias{SampleSize}
\title{SampleSize object}
\usage{
SampleSize(sample.size)
}
\arguments{
\item{sample.size}{a list or vector of sample size(s).}
}
\description{
This function creates an object of class \code{SampleSize} which can be
added to an object of class \code{DataModel}.
}
\details{
Objects of class \code{SampleSize} are used in objects of class
\code{DataModel} to specify the sample size in case of balanced design (all
samples will have the same sample size). A single object of class
\code{SampleSize} can be added to an object of class \code{DataModel}.

Either objects of class \code{Event} or \code{SampleSize} can be added to an
object of class \code{DataModel}, but not both.
}
\examples{

# Outcome parameter set 1
outcome1.placebo = parameters(mean = 0, sd = 70)
outcome1.treatment = parameters(mean = 40, sd = 70)

# Outcome parameter set 2
outcome2.placebo = parameters(mean = 0, sd = 70)
outcome2.treatment = parameters(mean = 50, sd = 70)

# Data model
case.study1.data.model = DataModel() +
                         OutcomeDist(outcome.dist = "NormalDist") +
                         SampleSize(c(50, 55, 60, 65, 70)) +
                         Sample(id = "Placebo",
                                outcome.par = parameters(outcome1.placebo, outcome2.placebo)) +
                         Sample(id = "Treatment",
                                outcome.par = parameters(outcome1.treatment, outcome2.treatment))

# Equivalent to:
case.study1.data.model = DataModel() +
                         OutcomeDist(outcome.dist = "NormalDist") +
                         SampleSize(seq(50, 70, 5)) +
                         Sample(id = "Placebo",
                                outcome.par = parameters(outcome1.placebo, outcome2.placebo)) +
                         Sample(id = "Treatment",
                                outcome.par = parameters(outcome1.treatment, outcome2.treatment))

}
\references{
\url{http://gpaux.github.io/Mediana/}
}
\seealso{
See Also \code{\link{DataModel}}.
}
