\name{families}
\alias{tests}
\alias{samples}
\alias{statistics}
\alias{parameters}
\alias{families}

\title{Create list of character strings
}
\description{
This function is used mostly for user's convenience. It simply creates a list of character strings.
}
\usage{
tests(...)
samples(...)
statistics(...)
parameters(...)
families(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
defines character strings to be passed into the function.
}
}

\references{
\url{http://gpaux.github.io/Mediana/}
}
